fix: incorrect mountpoint if devicelist changed order
If the devicelist function returned objects in a different order
the existing function could return an incorrect mountpoint path.

Fixed this by refactoring the method.

Additionally, it looks like this method will loop forever until
a valid device and mountpoint is found. Unsure if limiting this
to a maximum retry count would break things, so leaving as
unlimited retries.

Signed-off-by: Lawrence Brooks <84.lbrooks@gmail.com>